# docs/swagger.yaml
basePath: /dev/api/v1
definitions:
  controller._ResponseLoginData:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        default: 1005
      data:
        $ref: '#/definitions/models.ReqLogin'
      msg:
        default: success
        type: string
    type: object
  controller.ResCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 1011
    - 1012
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeEmailExist
    - CodeEmailNotExist
    - CodeInvalidPassword
    - CodeInvalidCaptcha
    - CodeCaptchaNotExistOrTimeOut
    - CodeUnbound
    - CodeServerBusy
    - CodeInvalidToken
    - CodeNeedLogin
    - CodeInvalidCookie
    - CodeNotData
  controller.ResponseData:
    properties:
      code:
        $ref: '#/definitions/controller.ResCode'
      data: {}
      msg: {}
    type: object
  models.JsonCourse:
    properties:
      category:
        type: string
      classId:
        type: string
      location:
        type: string
      method:
        type: string
      name:
        type: string
      section:
        type: integer
      sectionCount:
        type: integer
      teacher:
        type: string
      teachingClasses:
        type: string
      weekday:
        type: integer
      weeks:
        items:
          type: integer
        type: array
    type: object
  models.JsonGrades:
    properties:
      classid:
        type: string
      credits:
        description: 学分
        type: number
      fraction:
        type: number
      gpa:
        description: 绩点
        type: number
      grade:
        type: string
      gradepoints:
        description: 学分绩点
        type: number
      isdegree:
        type: boolean
      name:
        type: string
      teacher:
        type: string
    type: object
  models.ParamBind:
    properties:
      password:
        type: string
      studentID:
        type: string
    required:
    - password
    - studentID
    type: object
  models.ParamCourse:
    properties:
      cookie:
        type: string
      semester:
        enum:
        - 3
        - 12
        type: integer
      year:
        type: integer
    required:
    - semester
    - year
    type: object
  models.ParamFeedBack:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
  models.ParamGpa:
    properties:
      cookie:
        type: string
    type: object
  models.ParamGradeDetaile:
    properties:
      classid:
        type: string
      cookie:
        type: string
      semester:
        enum:
        - 3
        - 12
        type: integer
      year:
        type: integer
    required:
    - semester
    - year
    type: object
  models.ParamGrades:
    properties:
      cookie:
        type: string
      semester:
        enum:
        - 3
        - 12
        type: integer
      year:
        type: integer
    required:
    - semester
    - year
    type: object
  models.ParamLogin:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  models.ParamReCover:
    properties:
      captcha:
        type: string
      email:
        type: string
      newpassword:
        minLength: 8
        type: string
      renewpassword:
        minLength: 8
        type: string
    required:
    - captcha
    - email
    - newpassword
    - renewpassword
    type: object
  models.ParamReSet:
    properties:
      newpassword:
        minLength: 8
        type: string
      password:
        minLength: 8
        type: string
      renewpassword:
        minLength: 8
        type: string
    required:
    - newpassword
    - password
    - renewpassword
    type: object
  models.ParamRecord:
    properties:
      content:
        type: string
      id:
        type: string
      title:
        type: string
      type:
        enum:
        - 1
        - 2
        - 3
        type: integer
    required:
    - content
    - title
    - type
    type: object
  models.ParamSignUp:
    properties:
      captcha:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      repassword:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - captcha
    - email
    - password
    - repassword
    - username
    type: object
  models.ReqBind:
    properties:
      cookie:
        type: string
      syluinfo:
        $ref: '#/definitions/models.ReqSyluInfo'
    type: object
  models.ReqCourse:
    properties:
      courses:
        items:
          $ref: '#/definitions/models.JsonCourse'
        type: array
      starttime:
        type: string
    type: object
  models.ReqLogin:
    properties:
      email:
        type: string
      syluinfo:
        $ref: '#/definitions/models.ReqSyluInfo'
      token:
        type: string
      username:
        type: string
    type: object
  models.ReqSyluInfo:
    properties:
      college:
        default: 信息科学与工程学院
        description: 学院
        type: string
      grade:
        default: "2022"
        description: 年级
        type: string
      major:
        default: 计算机科学与技术(0305)
        description: 专业
        type: string
      reusername:
        default: 肖嘉兴
        description: 真实姓名
        type: string
      studentID:
        default: "2203050212"
        description: 学号
        type: string
    type: object
  models.ResGpa:
    properties:
      allgpa:
        type: string
      degreegpa:
        type: string
    type: object
  models.ResGradeDetail:
    properties:
      name:
        type: string
      score:
        type: string
      weight:
        type: string
    type: object
  models.ResSchoolCale:
    properties:
      endtime:
        type: string
      schoolcale:
        items:
          $ref: '#/definitions/models.SchoolCale'
        type: array
      starttime:
        type: string
      title:
        type: string
    type: object
  models.ResSemeSter:
    properties:
      index:
        type: integer
      list:
        items:
          $ref: '#/definitions/models.SemeSterList'
        type: array
    type: object
  models.SchoolCale:
    properties:
      endtime:
        type: string
      id:
        type: string
      name:
        type: string
      starttime:
        type: string
    type: object
  models.SemeSterList:
    properties:
      month:
        type: integer
      name:
        type: string
      year:
        type: integer
    type: object
info:
  contact:
    name: hakcerxiao
    url: http://www.hackerxiao.online
  description: 致力于为同学们提供校园服务(忽略状态码，所有响应都是200)
  title: we校园小助手项目接口文档
  version: "1.1"
paths:
  /auth/coskey:
    get:
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "1000":
          description: code=1000,msg="success",data里面是cos临时密钥数据
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "1005":
          description: code=1000+，msg里面是错误信息,data=null
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取COS临时密钥接口
      tags:
      - auth相关接口
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录参数,必填
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamLogin'
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",成功返回token,若绑定教务信息，也会包含教务学生信息
          schema:
            $ref: '#/definitions/controller._ResponseLoginData'
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 登录接口
      tags:
      - auth相关接口
  /auth/recoverpass:
    post:
      parameters:
      - description: 使用邮箱验证码新密码重置
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamReCover'
      responses:
        "1000":
          description: code=1000,msg="success",data=null
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "1005":
          description: code=1000+，msg里面是错误信息,data=null
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 找回密码验证接口
      tags:
      - auth相关接口
  /auth/resetpass:
    post:
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 使用旧密码新密码重置
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamReSet'
      responses:
        "1000":
          description: code=1000,msg="success",data=null
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "1005":
          description: code=1000+，msg里面是错误信息,data=null
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 重置密码验证接口
      tags:
      - auth相关接口
  /auth/sendemail:
    get:
      parameters:
      - in: query
        name: email
        required: true
        type: string
      - enum:
        - sign
        - recoverpass
        in: query
        name: mode
        required: true
        type: string
      responses:
        "1000":
          description: code=1000,msg="success",data=null
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "1005":
          description: code=1000+，msg里面是错误信息,data=null
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 邮箱验证接口
      tags:
      - auth相关接口
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册参数,必填
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamSignUp'
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",data=null
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "1001":
          description: code=1000+，msg里面是错误信息,data=null
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 注册接口
      tags:
      - auth相关接口
  /edu/bind:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 绑定参数,必填
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamBind'
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",
          schema:
            $ref: '#/definitions/models.ReqBind'
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 绑定接口
      tags:
      - 校园服务相关接口
  /edu/cale:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: query一个cookie即可
        in: query
        name: cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",
          schema:
            $ref: '#/definitions/models.ResSchoolCale'
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取校历接口
      tags:
      - 校园服务相关接口
  /edu/cookie:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",
          schema:
            type: string
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取cookie接口
      tags:
      - 校园服务相关接口
  /edu/courses:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 课表参数,必填,其中semester为3或12表示某学期，例如year=2022 semester=3 表示2022-2023学年第一学期
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamCourse'
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",
          schema:
            $ref: '#/definitions/models.ReqCourse'
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取课表接口
      tags:
      - 校园服务相关接口
  /edu/courses/auto:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: query一个cookie即可，自动推断当前学期
        in: query
        name: cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",
          schema:
            $ref: '#/definitions/models.ReqCourse'
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 自动获取课表接口
      tags:
      - 校园服务相关接口
  /edu/gpas:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 此接口响应时间>=4s，暂无优化思路
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamGpa'
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",
          schema:
            $ref: '#/definitions/models.ResGpa'
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取绩点接口
      tags:
      - 校园服务相关接口
  /edu/grade/detaile:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 成绩参数,必填，其中semester为3或12表示某学期，例如year=2022 semester=3 表示2022-2023学年第一学期
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamGradeDetaile'
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",
          schema:
            items:
              $ref: '#/definitions/models.ResGradeDetail'
            type: array
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取成绩详情接口
      tags:
      - 校园服务相关接口
  /edu/grades:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 成绩参数,必填，其中semester为3或12表示某学期，例如year=2022 semester=3 表示2022-2023学年第一学期
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamGrades'
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",
          schema:
            items:
              $ref: '#/definitions/models.JsonGrades'
            type: array
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取成绩接口
      tags:
      - 校园服务相关接口
  /edu/semester:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",
          schema:
            $ref: '#/definitions/models.ResSemeSter'
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取学期接口
      tags:
      - 校园服务相关接口
  /feedback:
    post:
      consumes:
      - application/json
      parameters:
      - description: 反馈参数,必填
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamFeedBack'
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",data="null
          schema:
            type: string
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 反馈接口
      tags:
      - 反馈相关接口
  /socail/record:
    post:
      consumes:
      - application/json
      parameters:
      - description: 反馈参数,必填
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamRecord'
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",data="null
          schema:
            type: string
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 便签curd
      tags:
      - 便签相关接口
  /socail/records:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "1000":
          description: code=1000,msg="success",data="null
          schema:
            type: string
        "1001":
          description: 请求错误参数,code=1000+，msg里面是错误信息
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取全部便签接口
      tags:
      - 便签相关接口
swagger: "2.0"